remove show() button
This stores a bit more state in the environment but allows just pressing
'next' to keep showing angles and their values.  This makes it slightly
nicer to use in that you don't need to keep switching which button
you're pressing and can just keep hitting the space bar to proceed.